+++ lib/qc/modules/d2c_video/components/d2c_video_home_component.g.dart
@@ -1,0 +1,29 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'd2c_video_home_component.dart';
+
+// **************************************************************************
+// ComponentGenerator
+// **************************************************************************
+
+dynamic schema() => {
+      //#admincomponent
+      "type": "@@component",
+      "key": "QC_d2c_video_home_component",
+      "componentType": "QC D2C Video Home Component",
+      "isActive": true,
+      "title": "D2c Video Home Component",
+      "configJson": {
+        "type": "list",
+        "config": [
+          {
+            "uiType": "input",
+            "type": "String",
+            "isRequired": false,
+            "label": "None",
+            "key": "none"
+          }
+        ]
+      }
+      //#admincomponent
+    };