+++ lib/qc/modules/d2c_video/screens/d2c_video_home_screen.g.dart
@@ -1,0 +1,18 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'd2c_video_home_screen.dart';
+
+// **************************************************************************
+// Generator: PageSchemaGen
+// **************************************************************************
+
+dynamic schema() => {
+      //#adminscreen
+      "type": "@@page",
+      "key": "QC_d2c_video_home_screen",
+      "pageGroup": "QC D2C Video Home Page",
+      "title": "D2c Video Home Screen",
+      "isActive": true,
+      "params": null
+      //#adminscreen
+    };