+++ lib/qc/qc_routes.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_trc/qc/modules/d2c_video/screens/d2c_lot_device_listing_screen.dart';
+import 'package:flutter_trc/qc/modules/d2c_video/screens/d2c_lot_listing_screen.dart';
 import 'package:flutter_trc/qc/modules/d2c_video/screens/d2c_video_screen.dart';
 import 'package:flutter_trc/qc/modules/data_wipe/screens/data_wipe_home_screen.dart';
 import 'package:flutter_trc/qc/modules/data_wipe/screens/data_wipe_list_screen.dart';
@@ -39,6 +41,7 @@ import 'package:flutter_trc/src/common/mpin/screens/mpin_setup_screen.dart';
 import 'package:flutter_trc/src/common/user/widget/user_profile_screen.dart';
 import 'package:flutter_trc/src/common/widgets/dispute_image_editor_screen.dart';
 
+import 'modules/d2c_video/screens/d2c_video_home_screen.dart';
 import 'modules/data_wipe/screens/data_wipe_detail_screen.dart';
 import 'modules/dead_repair/screens/index.dart';
 import 'modules/device_receive_module/screens/device_receive_screen.dart';
@@ -116,6 +119,9 @@ class QcRoutes {
       DataWipeDetailScreen.route: (_) => const DataWipeDetailScreen(),
       DataWipeListScreen.route: (_) => const DataWipeListScreen(),
       DataWipeHomeScreen.route: (_) => const DataWipeHomeScreen(),
+      D2cVideoHomeScreen.route: (_) => const D2cVideoHomeScreen(),
+      D2cLotListingScreen.route: (_) => const D2cLotListingScreen(),
+      D2cLotDeviceListingScreen.route: (_) => const D2cLotDeviceListingScreen(),
     };
   }
 }