+++ lib/src/common/mpin/l10n.dart
@@ -32,8 +32,13 @@ class L10n extends BaseL10n {
 
   String get enableFingerprint => Intl.message('Enable Fingerprint', locale: localName, name: 'enableFingerprint');
 
+  String get enableFaceId => Intl.message('Enable Face ID', locale: localName, name: 'enableFaceId');
+
   String get enableFingerprintDesc => Intl.message('With Touch ID, you won’t need to enter your mPIN every time.',
       locale: localName, name: 'enableFingerprintDesc');
+
+  String get enableFaceIdDesc => Intl.message('With Face ID, you won’t need to enter your mPIN every time.',
+      locale: localName, name: 'enableFaceIdDesc');
 
   String get enable => Intl.message('Enable', locale: localName, name: 'enable');
 
@@ -47,6 +52,9 @@ class L10n extends BaseL10n {
 
   String get loginUsingFingerprint =>
       Intl.message('Login Using Fingerprint', locale: localName, name: 'loginUsingFingerprint');
+
+  String get loginUsingFaceId =>
+      Intl.message('Login Using Face Id', locale: localName, name: 'loginUsingFaceId');
 
   String get registrationSuccessful =>
       Intl.message('Registration successful', locale: localName, name: 'registrationSuccessful');